Bug fixes and adjustments
fix bug where glitched in to wall after grabbing on to ledge. Removed timer signal connection and a couple of un-necessary functions. Adjusted ray casts and pick up bit audio. Changed crush area layer and turned monitorable on and set monitoring to false in die() to fix bug where not dying holding on to moving platforms.